Extract Line class in a different file
